+++ src/anonymizer/view/series.py
@@ -3,7 +3,6 @@ import logging
 import os
 import tkinter as ttk
 from pathlib import Path
-from pprint import pformat
 
 import customtkinter as ctk
 import numpy as np
@@ -12,7 +11,7 @@ from easyocr import Reader
 from pydicom import Dataset
 
 from anonymizer.controller.create_projections import load_series_frames
-from anonymizer.controller.remove_pixel_phi import detect_text
+from anonymizer.controller.remove_pixel_phi import OCRText, detect_text
 from anonymizer.model.anonymizer import AnonymizerModel
 from anonymizer.utils.translate import _
 from anonymizer.view.image import ImageViewer
@@ -33,6 +32,7 @@ class SeriesView(ctk.CTkToplevel):
         self._anon_model = anon_model
         self._series_dir = series_path
         self._ocr_reader = None  # only created if user clicks "Detect Text" button
+        self.detected_text: dict[int, list[OCRText]] = {}  # Store all detected text per frame
 
         self.protocol("WM_DELETE_WINDOW", self._on_cancel)
         self.bind("<Escape>", self._escape_keypress)
@@ -65,8 +65,8 @@ class SeriesView(ctk.CTkToplevel):
 
         # Whitelist entry:
         self.whitelist_entry = ctk.CTkEntry(list_frame)
-        self.whitelist_entry.grid(row=1, columnspan=2, sticky="ew")
         self.whitelist_entry.bind("<Return>", self.whitelist_button_clicked_or_entry_return)
+        self.whitelist_entry.grid(row=1, columnspan=2, sticky="ew")
 
         # Whitelist listbox & scrollbar:
         scrollbar = ttk.Scrollbar(list_frame, orient="vertical")
@@ -74,12 +74,11 @@ class SeriesView(ctk.CTkToplevel):
         scrollbar.config(command=self.whitelist.yview)
         self.whitelist.bind("<Delete>", self.whitelist_delete_keypressed)
         self.whitelist.bind("<BackSpace>", self.whitelist_delete_keypressed)
-
         self.whitelist.grid(row=2, column=0, sticky="nsew")
         scrollbar.grid(row=2, column=1, sticky="ns")
 
         # ImageViewer:
-        self.image_viewer = ImageViewer(self._sv_frame, self._frames)
+        self.image_viewer = ImageViewer(self._sv_frame, self._frames, add_to_whitelist_callback=self.add_to_whitelist)
         self.image_viewer.grid(row=0, column=1, sticky="nsew")
 
         # Control Frame:
@@ -99,22 +98,47 @@ class SeriesView(ctk.CTkToplevel):
         )
         self.remove_button.grid(row=0, column=1, padx=self.PAD, pady=self.PAD, sticky="w")
 
+        # Status label:
+        self.status_label = ctk.CTkLabel(self.control_frame, text="")
+        self.status_label.grid(row=0, column=2, padx=self.PAD, pady=self.PAD, sticky="ew")
+        self.control_frame.grid_columnconfigure(2, weight=1)  # Make status label expand.
+
         self.populate_listbox()
+
+    def _update_title(self):
+        if self._ds:
+            phi = self._anon_model.get_phi(self._ds.PatientID)
+            if phi:
+                title = (
+                    _("Series View for")
+                    + f" {phi.patient_name} PHI ID:{phi.patient_id} ANON ID: {self._ds.PatientID}"
+                    + f" {self._ds.get('SeriesDescription', '')} "
+                )
+                self.title(title)
+
+    def update_status(self, message: str):
+        """Updates the status label."""
+        self.status_label.configure(text=message)
+        self.status_label.update()
+
+    def add_to_whitelist(self, text: str):
+        whitelist = self.whitelist.get(0, ttk.END)
+        try:
+            ndx = whitelist.index(text)
+            # Already in whitelist, hightlight:
+            self.whitelist.select_set(ndx)
+        except ValueError:
+            # Not in whitelist, insert:
+            logger.info(f"Adding to whitelist: {text}")
+            self.whitelist.insert(0, text)
+
+    def insert_entry_into_whitelist(self):
+        new_item = self.whitelist_entry.get()
+        self.add_to_whitelist(new_item)
+        self.whitelist_entry.delete(0, ctk.END)
 
     def whitelist_button_clicked_or_entry_return(self, event):
         self.insert_entry_into_whitelist()
-
-    def in_whitelist(self, item: str) -> bool:
-        return all(self.whitelist.get(i) != item for i in range(self.whitelist.size()))
-
-    def insert_entry_into_whitelist(self):
-        item = self.whitelist_entry.get()
-        if self.in_whitelist(item):
-            return
-
-        self.whitelist.insert(0, item)
-        # self.whitelist.select_set(0, 0)
-        self.whitelist_entry.delete(0, ctk.END)
 
     def whitelist_delete_keypressed(self, event):
         selected_indices = self.whitelist.curselection()
@@ -148,34 +172,6 @@ class SeriesView(ctk.CTkToplevel):
         )
         return ds, np.concatenate([projections, series_frames], axis=0)
 
-    def _update_title(self):
-        if self._ds:
-            phi = self._anon_model.get_phi(self._ds.PatientID)
-            if phi:
-                title = (
-                    _("Series View for")
-                    + f" {phi.patient_name} PHI ID:{phi.patient_id} ANON ID: {self._ds.PatientID}"
-                    + f" {self._ds.get('SeriesDescription', '')} "
-                )
-                self.title(title)
-
-    def _escape_keypress(self, event):
-        logger.info("_escape_pressed")
-        self._on_cancel()
-
-    def _on_cancel(self):
-        logger.info("_on_cancel")
-        self.grab_release()
-        self.destroy()
-
-        if hasattr(self, "image_viewer") and self.image_viewer:
-            self.image_viewer.clear_cache()
-
-        self._frames = None
-        self._ds = None
-        self._projection = None
-        gc.collect()
-
     def initialise_ocr(self):
         # Once-off initialisation of easyocr.Reader (and underlying pytorch model):
         # if pytorch models not downloaded yet, they will be when Reader initializes
@@ -191,13 +187,62 @@ class SeriesView(ctk.CTkToplevel):
             logger.info(f"EasyOCR downloaded models: {os.listdir(model_dir)}")
 
         # Initialize the EasyOCR reader with the desired language(s), if models are not in model_dir, they will be downloaded
+        # TODO: use thread for this if models not downloaded yet, provide status updates during download:
+        # TODO: optimize so reader is only initialised once - reference passed by ProjectionView or IndexView or global?
         self._ocr_reader = Reader(
             lang_list=["en", "de", "fr", "es"],
+            gpu=False,
             model_storage_directory=model_dir,
-            verbose=True,
+            verbose=False,
         )
 
         logging.info("OCR Reader initialised successfully")
+
+    def filter_text_data(self, frame_index: int) -> list[OCRText]:
+        """Filters the text_data for a frame based on the whitelist."""
+        if frame_index not in self.detected_text:
+            return []
+
+        whitelist_set = set(item.upper().strip() for item in self.whitelist.get(0, "end"))
+        filtered_results = []
+
+        for ocr_text in self.detected_text[frame_index]:
+            if ocr_text.text and ocr_text.text.upper() not in whitelist_set:
+                filtered_results.append(ocr_text)
+        return filtered_results
+
+    def draw_text_overlay(self, frame_index: int):
+        """Draws text boxes on the overlay for the given frame, based on filtered text_data."""
+        filtered_text_data = self.filter_text_data(frame_index)  # Filter *before* drawing
+        self.image_viewer.set_text_overlay_data(frame_index, filtered_text_data)
+
+    def process_single_frame_ocr(self, frame_index: int):
+        """Performs OCR on a single frame, filters, and stores results."""
+        with torch.no_grad():
+            if self._ocr_reader:
+                results = detect_text(
+                    pixels=self.image_viewer.images[frame_index], ocr_reader=self._ocr_reader, draw_boxes_and_text=False
+                )
+                if results:
+                    logger.info(results)
+                    self.detected_text[frame_index] = results
+                    self.draw_text_overlay(frame_index)
+
+    def detect_text_current_frame(self):
+        """Detects text in the currently displayed frame."""
+        self.update_status(_("OCR on current frame..."))
+        self.process_single_frame_ocr(self.image_viewer.current_image_index)
+        self.update_status(_("OCR on current frame complete"))
+
+    def detect_text_for_series(self):
+        """Detects text in all frames of the series."""
+        self.update_status(_("Detecting text in all frames..."))
+        total_frames = self.image_viewer.num_images
+        for i in range(total_frames):
+            self.image_viewer.load_and_display_image(i)  # Display image.
+            self.process_single_frame_ocr(i)
+            self.update_status(_("OCR on frame: ") + str(i + 1))
+        self.update_status(_("OCR on all frames complete"))
 
     def detect_text_button_clicked(self):
         logger.info("Detecting text...")
@@ -209,14 +254,7 @@ class SeriesView(ctk.CTkToplevel):
         logger.info(f"Apple MPS (Metal) GPU Available: {torch.backends.mps.is_available()}")
         logger.info(f"CUDA GPU Available: {torch.cuda.is_available()}")
 
-        with torch.no_grad():
-            if self._ocr_reader:
-                ocr_list = detect_text(
-                    pixels=self.image_viewer.get_current_image(), ocr_reader=self._ocr_reader, draw_boxes=True
-                )
-                if ocr_list:
-                    logger.info(f"Detected Text:\n{pformat(ocr_list)}")
-                    self.image_viewer.refresh_current_image()
+        self.detect_text_current_frame()
 
     def remove_text_button_clicked(self):
         logger.info("Removing text...")
@@ -227,24 +265,51 @@ class SeriesView(ctk.CTkToplevel):
         # translate dictionary into all languages, copy to relevant locales as phi_whitelist.csv
         # TODO: provide user option to load and use this dictionary
         # TODO: consider modality specific dictionaries, context sensitive to series modality
-        self.whitelist.insert(ttk.END, "L")
-        self.whitelist.insert(ttk.END, "R")
-        self.whitelist.insert(ttk.END, "PORTABLE")
-        self.whitelist.insert(ttk.END, "LEFT")
-        self.whitelist.insert(ttk.END, "RIGHT")
-        self.whitelist.insert(ttk.END, "SUPINE")
-        self.whitelist.insert(ttk.END, "PRONE")
-        self.whitelist.insert(ttk.END, "LATERAL")
-        self.whitelist.insert(ttk.END, "ANTERIOR")
-        self.whitelist.insert(ttk.END, "POSTERIOR")
-        self.whitelist.insert(ttk.END, "DECUBITUS")
-        self.whitelist.insert(ttk.END, "ERECT")
-        self.whitelist.insert(ttk.END, "OBLIQUE")
-        self.whitelist.insert(ttk.END, "AXIAL")
-        self.whitelist.insert(ttk.END, "CORONAL")
-        self.whitelist.insert(ttk.END, "SAGITTAL")
-        self.whitelist.insert(ttk.END, "STANDING")
-        self.whitelist.insert(ttk.END, "SITTING")
-        self.whitelist.insert(ttk.END, "RECUMBENT")
-        self.whitelist.insert(ttk.END, "UPRIGHT")
-        self.whitelist.insert(ttk.END, "SEMI-UPRIGHT")
+        whitelist_items = [
+            "L",
+            "R",
+            "PORTABLE",
+            "LEFT",
+            "RIGHT",
+            "SUPINE",
+            "PRONE",
+            "LATERAL",
+            "ANTERIOR",
+            "POSTERIOR",
+            "DECUBITUS",
+            "ERECT",
+            "OBLIQUE",
+            "AXIAL",
+            "CORONAL",
+            "SAGITTAL",
+            "STANDING",
+            "SITTING",
+            "RECUMBENT",
+            "UPRIGHT",
+            "SEMI-UPRIGHT",
+        ]
+        for item in whitelist_items:
+            self.whitelist.insert(ttk.END, item)
+
+    def _escape_keypress(self, event):
+        logger.info("_escape_pressed")
+        self._on_cancel()
+
+    def _on_cancel(self):
+        logger.info("_on_cancel")
+        self.grab_release()
+        self.destroy()
+
+        if hasattr(self, "image_viewer") and self.image_viewer:
+            self.image_viewer.clear_cache()
+
+        if self._ocr_reader:
+            # Attempt to unload the model
+            del self._ocr_reader
+            self._ocr_reader = None
+            gc.collect()
+
+        self._frames = None
+        self._ds = None
+        self._projection = None
+        gc.collect()