+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [17.4.*]
+### Added
+- SeriesView and ImageViewer classes with text detection and removal using easyocr reader
+
+### Changed
+- Modifications to index and projection modules for memory management and launching SeriesView with ImageViewer
 
 ## [17.3.*] 
 ### Changed